Corrected getBoolean and added getUntypedBoolean
getBoolean was only ever returning true, regardless of the actual value
supplied. The getType call in getBoolean was also suppressing the
functionality suggested by the following switch statement. So I copied
the getBoolean functionality to getUntypeBoolean and removed the type
restriction, and in getBoolean I removed the switch statement while
correcting the value retrieval.